refactor: :recycle: Split up windowed and windowless state into separate modules
