+++ src/gpu/pdb_gpu/mod.rs
@@ -1,11 +1,15 @@
 //! Adapted Tutorial 10
 
-use std::iter;
-
 // use image::{ImageBuffer, Rgba};
 // use tracing::warn;
 use wgpu::util::DeviceExt;
-use winit::{dpi::PhysicalSize, window::Window};
+use winit::dpi::PhysicalSize;
+
+use camera::{Camera, CameraController, CameraUniform};
+use instance::{Instance, InstanceRaw, LightUniform};
+use model::Vertex;
+
+use crate::gpu::pdb_gpu::input::UnifiedEvent;
 
 pub mod camera;
 pub mod input;
@@ -14,13 +18,9 @@ pub mod model;
 pub mod resources;
 pub mod run_tui;
 pub mod run_windowed;
+pub mod state_windowed;
+pub mod state_windowless;
 pub mod texture;
-
-use camera::{Camera, CameraController, CameraUniform};
-use instance::{Instance, InstanceRaw, LightUniform};
-use model::{DrawLight, DrawModel, Vertex};
-
-use crate::gpu::pdb_gpu::input::UnifiedEvent;
 
 #[rustfmt::skip]
 pub const OPENGL_TO_WGPU_MATRIX: nalgebra::Matrix4<f32> = nalgebra::Matrix4::new(
@@ -93,66 +93,9 @@ fn create_render_pipeline(
 }
 
 pub trait InnerState {
-    fn size(&self) -> winit::dpi::PhysicalSize<u32>;
+    fn size(&self) -> PhysicalSize<u32>;
     fn format(&self) -> wgpu::TextureFormat;
-    fn resize(&mut self, new_size: winit::dpi::PhysicalSize<u32>, device: &wgpu::Device);
-}
-
-#[derive(Debug)]
-struct WindowedState {
-    window: Window,
-    surface: wgpu::Surface,
-    config: wgpu::SurfaceConfiguration,
-    size: winit::dpi::PhysicalSize<u32>,
-}
-
-impl WindowedState {
-    pub fn new(
-        window: Window,
-        surface: wgpu::Surface,
-        size: winit::dpi::PhysicalSize<u32>,
-        adapter: &wgpu::Adapter,
-        device: &wgpu::Device,
-    ) -> Self {
-        let surface_caps = surface.get_capabilities(adapter);
-        let surface_format = surface_caps
-            .formats
-            .iter()
-            .copied()
-            .find(|f| f.is_srgb())
-            .unwrap_or(surface_caps.formats[0]);
-        let config = wgpu::SurfaceConfiguration {
-            usage: wgpu::TextureUsages::RENDER_ATTACHMENT,
-            format: surface_format,
-            width: size.width,
-            height: size.height,
-            present_mode: surface_caps.present_modes[0],
-            alpha_mode: surface_caps.alpha_modes[0],
-            view_formats: vec![],
-        };
-        surface.configure(device, &config);
-        Self {
-            window,
-            surface,
-            config,
-            size,
-        }
-    }
-}
-
-impl InnerState for WindowedState {
-    fn size(&self) -> winit::dpi::PhysicalSize<u32> {
-        self.size
-    }
-    fn format(&self) -> wgpu::TextureFormat {
-        self.config.format
-    }
-    fn resize(&mut self, new_size: winit::dpi::PhysicalSize<u32>, device: &wgpu::Device) {
-        self.size = new_size;
-        self.config.width = new_size.width;
-        self.config.height = new_size.height;
-        self.surface.configure(device, &self.config);
-    }
+    fn resize(&mut self, new_size: PhysicalSize<u32>, device: &wgpu::Device);
 }
 
 #[derive(Debug)]
@@ -368,7 +311,7 @@ impl<IS: InnerState> State<IS> {
         }
     }
     /// Resize the canvas
-    fn resize(&mut self, new_size: winit::dpi::PhysicalSize<u32>) {
+    pub fn resize(&mut self, new_size: PhysicalSize<u32>) {
         if new_size.width > 0 && new_size.height > 0 {
             self.inner_state.resize(new_size, &self.device);
             self.camera.aspect =
@@ -381,7 +324,7 @@ impl<IS: InnerState> State<IS> {
             );
         }
     }
-    fn update(&mut self) {
+    pub fn update(&mut self) {
         self.camera_controller.update_camera(&mut self.camera);
         self.camera_uniform.update_view_proj(&self.camera);
         self.queue.write_buffer(
@@ -407,7 +350,7 @@ impl<IS: InnerState> State<IS> {
     }
     // TODO Consider moving this function outside of `State`, like the function for creating a render pipeline
     /// Create the devices needed for cases with or without a window
-    async fn create_adapter_device_queue(
+    pub async fn create_adapter_device_queue(
         surface_option: Option<&wgpu::Surface>,
         instance: &wgpu::Instance,
     ) -> (wgpu::Adapter, wgpu::Device, wgpu::Queue) {
@@ -435,356 +378,5 @@ impl<IS: InnerState> State<IS> {
     }
     fn input(&mut self, event: UnifiedEvent) -> bool {
         self.camera_controller.process_events(event)
-    }
-}
-
-impl State<WindowedState> {
-    async fn new(window: Window) -> Self {
-        let size = window.inner_size();
-
-        // The instance is a handle to our GPU
-        // BackendBit::PRIMARY => Vulkan + Metal + DX12 + Browser WebGPU
-        let instance = wgpu::Instance::new(wgpu::InstanceDescriptor {
-            backends: wgpu::Backends::PRIMARY,
-            ..Default::default()
-        });
-
-        // TODO In later verison of `wgpu` this is annotated with lifetime and no longer needs to be unsafe
-        // State owns the window so this should be safe.
-        let surface = unsafe { instance.create_surface(&window) }.unwrap();
-
-        let (adapter, device, queue) =
-            Self::create_adapter_device_queue(Some(&surface), &instance).await;
-        let inner_state = WindowedState::new(window, surface, size, &adapter, &device);
-        Self::new_from_inner_state(inner_state, device, queue).await
-    }
-    pub fn window(&self) -> &Window {
-        &self.inner_state.window
-    }
-
-    fn render(&mut self) -> Result<(), wgpu::SurfaceError> {
-        let output = self.inner_state.surface.get_current_texture()?;
-        let view = output
-            .texture
-            .create_view(&wgpu::TextureViewDescriptor::default());
-
-        let mut encoder = self
-            .device
-            .create_command_encoder(&wgpu::CommandEncoderDescriptor {
-                label: Some("Render Encoder"),
-            });
-
-        {
-            let mut render_pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-                label: Some("Render Pass"),
-                color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-                    view: &view,
-                    resolve_target: None,
-                    ops: wgpu::Operations {
-                        load: wgpu::LoadOp::Clear(wgpu::Color {
-                            r: 0.9,
-                            g: 0.9,
-                            b: 0.9,
-                            a: 1.0,
-                        }),
-                        store: wgpu::StoreOp::Store,
-                    },
-                })],
-                depth_stencil_attachment: Some(wgpu::RenderPassDepthStencilAttachment {
-                    view: &self.depth_texture.view,
-                    depth_ops: Some(wgpu::Operations {
-                        load: wgpu::LoadOp::Clear(1.0),
-                        store: wgpu::StoreOp::Store,
-                    }),
-                    stencil_ops: None,
-                }),
-                occlusion_query_set: None,
-                timestamp_writes: None,
-            });
-
-            render_pass.set_vertex_buffer(1, self.instance_buffer.slice(..));
-            render_pass.set_pipeline(&self.light_render_pipeline);
-            render_pass.draw_light_model(
-                &self.obj_model,
-                &self.camera_bind_group,
-                &self.light_bind_group,
-            );
-            render_pass.set_pipeline(&self.render_pipeline);
-            render_pass.draw_model_instanced(
-                &self.obj_model,
-                0..self.instances.len() as u32,
-                &self.camera_bind_group,
-                &self.light_bind_group,
-            );
-        }
-
-        self.queue.submit(iter::once(encoder.finish()));
-        output.present();
-
-        Ok(())
-    }
-}
-
-#[derive(Debug)]
-struct WindowlessState {
-    size: winit::dpi::PhysicalSize<u32>,
-    output_buffer: wgpu::Buffer,
-    output_image: Vec<u8>,
-    texture: wgpu::Texture,
-}
-
-impl WindowlessState {
-    const U32_SIZE: u32 = std::mem::size_of::<u32>() as u32;
-    const FORMAT: wgpu::TextureFormat = wgpu::TextureFormat::Rgba8UnormSrgb;
-
-    // Take a number of bytes and return the next closest multiple of 256
-    pub fn pad_bytes_to_256(bytes: u32) -> u32 {
-        (bytes + 255) & !255
-    }
-
-    // Pad width to 64 since each pixel requires 4 bytes
-    pub fn pad_width_to_64(width: u32) -> u32 {
-        (width + 63) & !63
-    }
-
-    pub fn new(size: winit::dpi::PhysicalSize<u32>, device: &wgpu::Device) -> Self {
-        // TODO Need to add functionality for changing this
-        let output_buffer_size = (Self::U32_SIZE * Self::pad_width_to_64(size.width) * size.height)
-            as wgpu::BufferAddress;
-        let output_buffer_desc = wgpu::BufferDescriptor {
-            size: output_buffer_size,
-            usage: wgpu::BufferUsages::COPY_DST | wgpu::BufferUsages::MAP_READ,
-            label: Some("Windowless Output Buffer"),
-            mapped_at_creation: false,
-        };
-        let output_buffer = device.create_buffer(&output_buffer_desc);
-
-        let texture_desc = wgpu::TextureDescriptor {
-            size: wgpu::Extent3d {
-                width: size.width,
-                height: size.height,
-                depth_or_array_layers: 1,
-            },
-            mip_level_count: 1,
-            sample_count: 1,
-            dimension: wgpu::TextureDimension::D2,
-            format: Self::FORMAT,
-            view_formats: &[], // NOTE This may be incorrect and needs to be checked
-            usage: wgpu::TextureUsages::COPY_SRC | wgpu::TextureUsages::RENDER_ATTACHMENT,
-            label: Some("Windowless Output Texture"),
-        };
-        let texture = device.create_texture(&texture_desc);
-
-        // Multiply by 4 because RGBA
-        let output_image_size = size.width as usize * size.height as usize * 4;
-        let output_image = Vec::<u8>::with_capacity(output_image_size);
-
-        Self {
-            size,
-            output_buffer,
-            output_image,
-            texture,
-        }
-    }
-}
-
-impl InnerState for WindowlessState {
-    fn size(&self) -> winit::dpi::PhysicalSize<u32> {
-        self.size
-    }
-    fn format(&self) -> wgpu::TextureFormat {
-        wgpu::TextureFormat::Rgba8UnormSrgb
-    }
-    fn resize(&mut self, new_size: winit::dpi::PhysicalSize<u32>, device: &wgpu::Device) {
-        self.size = new_size;
-
-        self.output_buffer.destroy();
-        self.texture.destroy();
-
-        // TODO Find a solution without repeating so much code
-        let output_buffer_size = (Self::U32_SIZE
-            * Self::pad_width_to_64(self.size.width)
-            * self.size.height) as wgpu::BufferAddress;
-        let output_buffer_desc = wgpu::BufferDescriptor {
-            size: output_buffer_size,
-            usage: wgpu::BufferUsages::COPY_DST | wgpu::BufferUsages::MAP_READ,
-            label: Some("Windowless Output Buffer"),
-            mapped_at_creation: false,
-        };
-        self.output_buffer = device.create_buffer(&output_buffer_desc);
-
-        // TODO Also recreate the texture
-        let texture_desc = wgpu::TextureDescriptor {
-            size: wgpu::Extent3d {
-                width: self.size.width,
-                height: self.size.height,
-                depth_or_array_layers: 1,
-            },
-            mip_level_count: 1,
-            sample_count: 1,
-            dimension: wgpu::TextureDimension::D2,
-            format: Self::FORMAT,
-            view_formats: &[], // NOTE This may be incorrect and needs to be checked
-            usage: wgpu::TextureUsages::COPY_SRC | wgpu::TextureUsages::RENDER_ATTACHMENT,
-            label: Some("Windowless Output Texture"),
-        };
-        self.texture = device.create_texture(&texture_desc);
-
-        // TODO Work out logic for new offset
-    }
-}
-
-impl State<WindowlessState> {
-    async fn new(size: PhysicalSize<u32>) -> Self {
-        // The instance is a handle to our GPU
-        // BackendBit::PRIMARY => Vulkan + Metal + DX12 + Browser WebGPU
-        let instance = wgpu::Instance::new(wgpu::InstanceDescriptor {
-            backends: wgpu::Backends::PRIMARY,
-            ..Default::default()
-        });
-
-        let (_adapter, device, queue) = Self::create_adapter_device_queue(None, &instance).await;
-        let inner_state = WindowlessState::new(size, &device);
-        let mut state = Self::new_from_inner_state(inner_state, device, queue).await;
-
-        // Accounting for the fact that font height is roughly twice the width
-        state.camera.aspect /= 2.0;
-        state.update();
-        state
-    }
-    // TODO Need to change this error
-    // TODO Need to refactor more out of this function
-    async fn render(&mut self) -> Result<(), wgpu::SurfaceError> {
-        let texture_view = self
-            .inner_state
-            .texture
-            .create_view(&wgpu::TextureViewDescriptor::default());
-
-        let mut encoder = self
-            .device
-            .create_command_encoder(&wgpu::CommandEncoderDescriptor {
-                label: Some("Render Encoder"),
-            });
-
-        {
-            let mut render_pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-                label: Some("Render Pass"),
-                color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-                    view: &texture_view,
-                    resolve_target: None,
-                    ops: wgpu::Operations {
-                        load: wgpu::LoadOp::Clear(wgpu::Color {
-                            r: 0.9,
-                            g: 0.9,
-                            b: 0.9,
-                            a: 1.0,
-                        }),
-                        store: wgpu::StoreOp::Store,
-                    },
-                })],
-                depth_stencil_attachment: Some(wgpu::RenderPassDepthStencilAttachment {
-                    view: &self.depth_texture.view,
-                    depth_ops: Some(wgpu::Operations {
-                        load: wgpu::LoadOp::Clear(1.0),
-                        store: wgpu::StoreOp::Store,
-                    }),
-                    stencil_ops: None,
-                }),
-                occlusion_query_set: None,
-                timestamp_writes: None,
-            });
-
-            render_pass.set_vertex_buffer(1, self.instance_buffer.slice(..));
-            render_pass.set_pipeline(&self.light_render_pipeline);
-            render_pass.draw_light_model(
-                &self.obj_model,
-                &self.camera_bind_group,
-                &self.light_bind_group,
-            );
-            render_pass.set_pipeline(&self.render_pipeline);
-            render_pass.draw_model_instanced(
-                &self.obj_model,
-                0..self.instances.len() as u32,
-                &self.camera_bind_group,
-                &self.light_bind_group,
-            );
-        }
-
-        encoder.copy_texture_to_buffer(
-            wgpu::ImageCopyTexture {
-                aspect: wgpu::TextureAspect::All,
-                texture: &self.inner_state.texture,
-                mip_level: 0,
-                origin: wgpu::Origin3d::ZERO,
-            },
-            wgpu::ImageCopyBuffer {
-                buffer: &self.inner_state.output_buffer,
-                layout: wgpu::ImageDataLayout {
-                    offset: 0,
-                    // Check that this isn't mean to be 4 `u8`s rather than 1 `u32`
-                    bytes_per_row: Some({
-                        let bytes = WindowlessState::U32_SIZE * self.inner_state.size().width;
-                        WindowlessState::pad_bytes_to_256(bytes)
-                    }),
-                    rows_per_image: Some(self.inner_state.size().height),
-                },
-            },
-            // TODO Stop redefining the same size
-            wgpu::Extent3d {
-                width: self.inner_state.size().width,
-                height: self.inner_state.size().height,
-                depth_or_array_layers: 1,
-            },
-        );
-
-        self.queue.submit(iter::once(encoder.finish()));
-
-        let buffer_slice = self.inner_state.output_buffer.slice(..);
-
-        // NOTE: We have to create the mapping THEN device.poll() before await
-        // the future. Otherwise the application will freeze.
-        let (tx, rx) = flume::bounded(1);
-        buffer_slice.map_async(wgpu::MapMode::Read, move |result| {
-            tx.send(result).unwrap();
-        });
-        self.device.poll(wgpu::Maintain::Wait);
-        rx.recv_async().await.unwrap().unwrap();
-
-        {
-            let data = buffer_slice.get_mapped_range();
-            self.inner_state.output_image.clear();
-            self.inner_state.output_image.extend_from_slice(&data[..]);
-        }
-
-        self.inner_state.output_buffer.unmap();
-
-        self.inner_state.output_image = self
-            .inner_state
-            .output_image
-            .chunks(
-                WindowlessState::U32_SIZE as usize
-                    * WindowlessState::pad_width_to_64(self.inner_state.size().width) as usize,
-            )
-            .flat_map(|row| {
-                row.iter().take(
-                    WindowlessState::U32_SIZE as usize * self.inner_state.size().width as usize,
-                )
-            })
-            .cloned()
-            .collect();
-
-        // TODO Fix the strangely sized buffer
-        // let now = chrono::Utc::now();
-        // let now_string = now.format("%H:%M:%S").to_string();
-        // let path = format!("From_inner_state_{}.png", now_string);
-        // let buffer = ImageBuffer::<Rgba<u8>, _>::from_raw(
-        //     WindowlessState::pad_width_to_64(self.inner_state.size().width),
-        //     self.inner_state.size().height,
-        //     &self.inner_state.output_image[..],
-        // )
-        // .unwrap();
-        // buffer.save(path).unwrap();
-
-        Ok(())
     }
 }
