+++ src/gpu/pdb_gpu/run_tui.rs
@@ -5,7 +5,8 @@ use tracing_subscriber;
 use winit::dpi::PhysicalSize;
 
 use crate::gpu::pdb_gpu::input::{UnifiedEvent, UnifiedKeyCode};
-use crate::gpu::pdb_gpu::{InnerState, State, WindowlessState};
+use crate::gpu::pdb_gpu::state_windowless::WindowlessState;
+use crate::gpu::pdb_gpu::{InnerState, State};
 
 use crate::basic_rasterizer::BasicAsciiRasterizer;
 use crate::rasterizer::ColoredPixel;
